==> Score/01_SCRAPE_NAVER_COMMENT.ipynb <==
import requests
from bs4 import BeautifulSoup
import numpy as np
import pandas as pd
from selenium import webdriver

url_main = 'https://comic.naver.com/' # webtoon/finish 완결, webtoon/detail?titleId=xxx&no=1 1화
driver_path = 'data/chromedriver.exe'

# 19세 이상 웹툰 데이터 수집을 위한 selenium 사용
driver = webdriver.Chrome(executable_path=driver_path)

driver.get(url_main + 'webtoon/finish')

직접 로그인 진행

# 웹툰 상세 페이지 이동

r = requests.get(url_main + 'webtoon/finish').text
soup = BeautifulSoup(r, 'lxml')

a = soup.select('div.thumb a')[0]
a

a.attrs['href']

stars = soup.select('div.rating_type strong')
stars[:10]

stars = [float(star.get_text()) for star in stars]
stars[:10]

def get_webtoon_titleId(url_main=url_main):
    r = requests.get(url_main + 'webtoon/finish').text
    soup = BeautifulSoup(r, 'lxml')
    a_list = soup.select('div.thumb a')
    epi_url_list = []
    comment_url_list = []
    
    for a in a_list:
        epi_url = url_main + 'webtoon/detail' + a.attrs['href'].split('list')[-1]
        comment_url = url_main + 'comment/comment' + a.attrs['href'].split('list')[-1]
        epi_url_list.append(epi_url)
        comment_url_list.append(comment_url)
        
    return epi_url_list, comment_url_list

epi_url_list, comment_url_list = get_webtoon_titleId()

epi_url_list[:10]

comment_url_list[:10]

driver.get(epi_url_list[0] + '&no=1')

# 회차별 댓글 수집
- 2013년 5월 이후 업로드된 작품만 수집(베댓 시스템 생긴 시점)

## 제목, 날짜 수집

soup = BeautifulSoup(driver.page_source, 'html.parser')
soup.select('.date')

soup.select('.date')[0].get_text()

import datetime

date = soup.select('.date')[0].get_text()
dt_test = datetime.date(int(date.split('.')[0]), int(date.split('.')[1]), int(date.split('.')[2]))
dt_test

dt_test > datetime.date(2013, 5, 1)

soup.select('.detail')[0].get_text().split('\t')[0].strip()

## 댓글 수집

driver.get(comment_url_list[0] + '&no=1')

soup = BeautifulSoup(driver.page_source, 'html.parser')
soup.select('.u_cbox_contents')

soup.select('.u_cbox_cnt_recomm') # 베댓 추천 수

soup.select('.u_cbox_cnt_unrecomm') # 베댓 비추 수

# 베스트 댓글에서 전체 댓글로 변경
driver.find_element_by_css_selector('#cbox_module_wai_u_cbox_sort_option_tab2 > span.u_cbox_sort_label').click()

soup = BeautifulSoup(driver.page_source, 'html.parser')
soup.select('.u_cbox_contents')

soup.select('.u_cbox_count') # 전체 댓글 수

driver.get(epi_url_list[0] + '&no=1')

soup = BeautifulSoup(driver.page_source, 'html.parser')
soup.select('.pointTotalPerson em') # 위 아래에 있어서 2개 출력 (평가한 인원 수)

soup.select('#topPointTotalNumber')[0].get_text() # 평점

## 함수 생성

import time

basedate = datetime.date(2013, 5, 1)

def get_comment_epi(epi_url, comment_url, star, after=basedate, no=1):
    """
    epi_url : 회차 base url (작품 제목과 날짜를 가져오기 위한 url)
    comment_url : 회차 댓글 url (베스트 댓글 수 등을 가져오기 위한 url)
    star : 평균 평점 (예측을 위한 값(y))
    after : 베댓 시스템이 생긴 2013년 5월 이후로만 수집
    no : 회차
    """
    
    # 우선 날짜를 추출 하여 2013년 5월 전에 올라온 회차면 pass
    driver.get(epi_url + '&no={}'.format(no))
    time.sleep(0.5)
    soup = BeautifulSoup(driver.page_source, 'html.parser')
    date = soup.select('.date')[0].get_text()
    date = datetime.date(int(date.split('.')[0]), int(date.split('.')[1]), int(date.split('.')[2]))
    
    if date < basedate:
        return
    
    # 제목과 평점, 평가한 인원 수 가져오기
    title = soup.select('.detail')[0].get_text().split('\t')[0].strip()
    score = float(soup.select('#topPointTotalNumber')[0].get_text())
    score_num = int(soup.select('.pointTotalPerson em')[0].get_text())
    
    # 베스트 댓글 수, 내용, 추천 합, 비추천 합 추출
    driver.get(comment_url + '&no={}'.format(no))
    time.sleep(0.5)
    soup = BeautifulSoup(driver.page_source, 'html.parser')
    best_comments = soup.select('.u_cbox_contents')
    
    best_comments_num = len(best_comments)
    best_comments = [comment.get_text() for comment in best_comments]
    best_comments = '\n'.join(best_comments) # 댓글을 하나로 합침
    
    recommen = soup.select('.u_cbox_cnt_recomm')
    recommen = [int(num.get_text()) for num in recommen]
    sum_recommen = sum(recommen)
    
    unrecomm = soup.select('.u_cbox_cnt_unrecomm')
    unrecomm = [int(num.get_text()) for num in unrecomm]
    sum_unrecomm = sum(unrecomm)
    
    # 전체 댓글 수, 최근 15개 댓글 내용 추출
    comments_num = int(soup.select('.u_cbox_count')[0].get_text().replace(',', ''))
    
    driver.find_element_by_css_selector('#cbox_module_wai_u_cbox_sort_option_tab2 > span.u_cbox_sort_label').click()
    time.sleep(0.5)
    
    soup = BeautifulSoup(driver.page_source, 'html.parser')
    comments_ = soup.select('.u_cbox_contents')
    comments_ = [comment.get_text() for comment in comments_]
    comments_ = '\n'.join(comments_)
    
    return [title, score, score_num, best_comments_num, best_comments, sum_recommen, sum_unrecomm, comments_num, comments_, star]

# 테스트
result = get_comment_epi(epi_url_list[0], comment_url_list[0], stars[0])
result

# 테스트2 (2012년 작품)
result = get_comment_epi('https://comic.naver.com/webtoon/detail?titleId=318995', 'https://comic.naver.com/comment/comment?titleId=318995', 0)
result

type(result)

# 코드 돌리기 전에 반드시 클린봇 OFF로!!

import pandas as pd
from tqdm.auto import tqdm

results = []
for epi_url, comment_url, star in tqdm(zip(epi_url_list, comment_url_list, stars), total=len(epi_url_list)):
    result = get_comment_epi(epi_url, comment_url, star)
    if result is None:
        continue
    results.append(result)
    
df = pd.DataFrame(results, columns=['title', 'score', 'score_num', 'best_num', 'best', 'best_recomm', 'best_unrecomm', 'comment_num', 'comment', 'star'])
df.head()

df.to_csv('data/naver_comment.csv', encoding='utf8', index=False)

df.to_csv('data/naver_comment3.csv', encoding='utf-8-sig', index=False)

driver.close()
