==> Score/02_MODEL1_TFIDF.ipynb <==
# 필요 라이브러리 호출
import numpy as np
import pandas as pd
from konlpy.tag import Okt
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from lightgbm import LGBMRegressor
from xgboost import XGBRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error, mean_squared_error
import warnings
warnings.filterwarnings('ignore')

# 데이터 불러오기
df = pd.read_csv('data/naver_comment.csv', encoding='utf8')
df.head()

okt = Okt()
def okt_tokenizer(text):
    tokens_ko = okt.morphs(text, stem=True)
    return tokens_ko

import re

def hangul(text):
    return re.sub("[^가-힣ㄱ-하-ㅣ\\s]", "", text)

def train_to_evaluate(X, y, min_df, max_df):
    tfidf_vect = TfidfVectorizer(tokenizer=okt_tokenizer, ngram_range=(1, 2), min_df=min_df, max_df=max_df)
    tfidf_vect.fit(X)
    tfidf_matrix = tfidf_vect.transform(X)
    
    X_train, X_test, y_train, y_test = train_test_split(tfidf_matrix, y, test_size=0.2, random_state=0)
    X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)
    
    lgbm = LGBMRegressor()
    xgb = XGBRegressor()
    rf = RandomForestRegressor()

    regs = [lgbm, xgb, rf]

    result = []
    for reg in regs:
        reg.fit(X_train, y_train)
        pred = reg.predict(X_val)
        mae = mean_absolute_error(y_val, pred)
        rmse = np.sqrt(mean_squared_error(y_val, pred))
        print(reg.__class__.__name__)
        print(f'MAE : {mae:.4f}, RMSE : {rmse:.4f}')
        
    return regs

# 1. 베댓 + 최근 댓글 15개 통합

## 1-1. `min_df=0, max_df=0.9`, 특수문자도 함께 처리

%%time

X = df['best'] + '\n' + df['comment']
y = df['star']

regs1_1 = train_to_evaluate(X, y, 0, 0.9)

## 1-2. `min_df=3, max_df=0.95`, 특수문자도 함께 처리

%%time

regs1_2 = train_to_evaluate(X, y, 3, 0.95)

## 1-3. `min_df=0, max_df=0.9`, 특수문자 제거

%%time

regs1_3 = train_to_evaluate(X.apply(hangul), y, 0, 0.9)

## 1-4. `min_df=3, max_df=0.95`, 특수문자 제거

%%time

regs1_4 = train_to_evaluate(X.apply(hangul), y, 3, 0.95)

# 2. 최근 댓글 15개로만

## 2-1. `min_df=0, max_df=0.9`, 특수문자도 함께 처리

%%time

X = df['comment']
y = df['star']

regs2_1 = train_to_evaluate(X, y, 0, 0.9)

## 2-2. `min_df=3, max_df=0.95`, 특수문자도 함께 처리

regs2_2 = train_to_evaluate(X, y, 3, 0.95)

## 2-3. `min_df=0, max_df=0.9`, 특수문자 제거

%%time

regs2_3 = train_to_evaluate(X.apply(hangul), y, 0, 0.9)

## 2-4. `min_df=3, max_df=0.95`, 특수문자 제거

%%time

regs2_4 = train_to_evaluate(X.apply(hangul), y, 3, 0.95)

- 특수문자 제거를 하지 않는 것이 더 높은 예측력(아주 약간)을 가짐
- `min_df`와 `max_df`는 뭘로 하든지 비슷하다
- TF-IDF로 예측 결과 MAE는 0.22\~0.30, RMSE는 0.40\~0.55

# 3. 베댓으로만

## 3-1. `min_df=0, max_df=0.9`, 특수문자도 함께 처리

%%time

X = df['best']
y = df['star']

regs3_1 = train_to_evaluate(X, y, 0, 0.9)

## 3-2. `min_df=3, max_df=0.95`, 특수문자도 함께 처리

%%time

regs3_2 = train_to_evaluate(X, y, 3, 0.95)

## 3-3. `min_df=0, max_df=0.9`, 특수문자 제거

%%time

regs3_3 = train_to_evaluate(X.apply(hangul), y, 0, 0.9)

## 3-4. `min_df=3, max_df=0.95`, 특수문자 제거

%%time

regs3_4 = train_to_evaluate(X.apply(hangul), y, 3, 0.95)

오차의 크기가 베댓만 > 베댓+최신댓글 15개 > 최신 댓글 15개만 순이다.  
다시 말해 완결난 작품에 한해서 최신 댓글 15개가 베댓보다 더 평점 설명을 잘 한다는 뜻.  
바꿔 말하면 1화까지 와서 '지뢰작'임을 알려주는 댓글이 생각보다 있다 + 1화가 올라온 시점에서는 욕 대신 칭찬이 많았을 것임을 추측할 수 있다.
