==> Score/02_MODEL2_텍스트제외_Classification.ipynb <==
# 필요 라이브러리 호출
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from lightgbm import LGBMClassifier
from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
import warnings
warnings.filterwarnings('ignore')

# 데이터 불러오기
df = pd.read_csv('data/naver_comment_2.csv', encoding='utf-8-sig')
df.head()

def train_to_evaluate(X, y):
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)
    X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)
    
    lgbm = LGBMClassifier()
    xgb = XGBClassifier()
    rf = RandomForestClassifier()
    lr = LogisticRegression()

    regs = [lgbm, xgb, rf, lr]

    result = []
    for reg in regs:
        reg.fit(X_train, y_train)
        pred = reg.predict(X_val)
        acc = accuracy_score(y_val, pred)
        print(reg.__class__.__name__)
        print(f'Accuracy: {acc:.4f}')
        
    return regs

# 9.9 이상, 미만 비율이 6:4
def category(star):
    if star >= 9.9:
        return 1
    else:
        return 0

# 1. 텍스트 제외 모든 feature

X = df.drop(columns=['title', 'best', 'comment', 'star'])
y = df['star'].apply(category)

regs1 = train_to_evaluate(X, y)

# 2. 베댓 추천 수, 비추천 수 제외

X = df.drop(columns=['title', 'best', 'comment', 'star', 'best_recomm', 'best_unrecomm'])

regs2 = train_to_evaluate(X, y)

## 추가적으로 베댓 수도 제외

X = X.drop(columns=['best_num'])

regs2_a = train_to_evaluate(X, y)

## 베댓 수만 제외

X = df.drop(columns=['title', 'best', 'comment', 'star', 'best_num'])

regs2_b = train_to_evaluate(X, y)

# 3. 1화 평점 제외

X = df.drop(columns=['title', 'best', 'comment', 'star', 'score'])

regs3 = train_to_evaluate(X, y)

# 4. 1화 평점, 평가 인원으로만

X = df[['score', 'score_num']]

regs4 = train_to_evaluate(X, y)

당연한 결과겠지만, 1화 평점이 평균 평점을 예측하는 데 있어 매우 중요한 요소이며 그렇다고 1화 평점만으로는 평균 평점을 예측하는 데 무리가 있다.  
그리고 베댓 수, 베댓 추천 수, 베댓 비추천 수는 오히려 오차를 늘리는 요소이므로 향후 최종 모델에서는 제외하는 것으로...  
또한 TF-IDF 결과보다 더 우수한 결과가 나왔다.
