==> Score/02_MODEL3_EMBEDDING.ipynb <==
# 필요 라이브러리 호출
import numpy as np
import pandas as pd
from konlpy.tag import Okt
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.preprocessing.text import Tokenizer
import warnings
warnings.filterwarnings('ignore')

# 데이터 불러오기
df = pd.read_csv('data/naver_comment.csv', encoding='utf8')
df.head()

okt = Okt()
def okt_tokenizer(text):
    tokens_ko = okt.morphs(text, stem=True)
    return tokens_ko

import re

def hangul(text):
    return re.sub("[^가-힣ㄱ-하-ㅣ\\s]", "", text)

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Flatten, Dense, Embedding
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint

def train_to_evaluate(X, y, file_name, max_words=10000, maxlen=30, embedding_dim=100, *args):
    
    tokenizer = Tokenizer(num_words=max_words)
    tokenizer.fit_on_texts(X)
    X_seq = tokenizer.texts_to_sequences(X)

    X_pad = pad_sequences(X_seq, maxlen=maxlen)
    y = np.asarray(y)
    
    X_train, X_test, y_train, y_test = train_test_split(X_pad, y, test_size=0.2, random_state=0)
    X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)
    
    model = Sequential()
    model.add(Embedding(max_words, embedding_dim, input_length=maxlen))
    model.add(Flatten())
    for unit in args:
        model.add(Dense(unit, activation='relu'))
    model.add(Dense(1))
    
    model.compile(optimizer='adam', loss='mse', metrics=['mae'])
    
    cp = ModelCheckpoint(filepath='data/'+file_name,
                         monitor='val_loss',
                         save_best_only=True)
    
    es = EarlyStopping(monitor='val_loss',
                       patience=15)
    
    
    hist = model.fit(X_train, y_train,
                     epochs=100, batch_size=32,
                     validation_data=(X_val, y_val),
                     callbacks=[cp, es],
                     verbose=0)
    
    
    model.load_weights('data/'+file_name)
    mse, mae = model.evaluate(X_val, y_val, verbose=0)
#     print('MAE :', np.round(mae, 4), 'RMSE :', np.round(np.sqrt(mse), 4))
        
    return model, hist, mse, mae

# 1. 베댓 + 최근 댓글

X1 = df['best'] + '\n' + df['comment']
y = df['star']

X1 = X1.apply(okt_tokenizer).values

max([len(x) for x in X1])

min([len(x) for x in X1])

## IMDB로 연습할 때와 동일한 모델로 학습
- `num_words=10000`, `maxlen=30`, `embedding_dim=100`, `Dense(32)`
- 결과는 매우 좋지 않음(데이터 수가 적어서, 토큰화 된 단어의 수가 적어서, 영어와 달리 한국어는 적은 단어만 사용돼서?)

model1, hist1, mse1, mae1 = train_to_evaluate(X1, y, 'embedding1.h5', 10000, 30, 100, 32)

print('MAE :', np.round(mae1, 4), 'RMSE :', np.round(np.sqrt(mse1), 4))

## 파라미터 변경
- `num_words=30`, `maxlen=20`, `embedding_dim=30`, `Dense` 없음
- TF-IDF 변환 후 LGBM/XGBoost/RandomForest로 학습한 때와 비슷한 수준의 결과가 나옴

model1, hist1, mse1, mae1 = train_to_evaluate(X1, y, 'embedding1.h5', 30, 20, 30)

print('MAE :', np.round(mae1, 4), 'RMSE :', np.round(np.sqrt(mse1), 4))

파라미터 값이 작아야 좋은 결과가 나오는 것 같기 때문에 최적 파라미터를 모든 경우의 수를 적용해서 찾아봄

from tqdm import tqdm

bestmodel = None
bestloss = 9999
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'unit', 'mse', 'mae'])

for mxwords in tqdm(range(20, 200+1, 20)):
    for mxlen in [20, 30]:
        for embdim in range(25, 100+1, 25):
            for unit in [5, 16, 32, 48, 64]:
                model = train_to_evaluate(X1, y, 'embedding1.h5', mxwords, mxlen, embdim, unit)
                result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, unit, model[2], model[3]]
                if bestloss > model[2]:
                    bestmodel = model
                    bestloss = model[2]
                    bestmxwords = mxwords
                    bestmxlen = mxlen
                    bestembdim = embdim
                    bestunit = unit

print(f'MAE: {bestmodel[3]:.4f}, RMSE: {np.sqrt(bestmodel[2]):.4f}')
print(f'--bestmxwords: {bestmxwords}', f'--bestmxlen: {bestmxlen}', f'--bestembdim: {bestembdim}', f'--bestunit: {bestunit}', sep='\t')

## 특문 제거

X1_a = df['best'].apply(hangul)
X1_a = X1_a.apply(okt_tokenizer).values

bestmodel = None
bestloss = 9999
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'unit', 'mse', 'mae'])

for mxwords in tqdm(range(20, 200+1, 20)):
    for mxlen in [20, 30]:
        for embdim in range(25, 100+1, 25):
            for unit in [5, 16, 32, 48, 64]:
                model = train_to_evaluate(X1_a, y, 'embedding1.h5', mxwords, mxlen, embdim, unit)
                result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, unit, model[2], model[3]]
                if bestloss > model[2]:
                    bestmodel = model
                    bestloss = model[2]
                    bestmxwords = mxwords
                    bestmxlen = mxlen
                    bestembdim = embdim
                    bestunit = unit

print(f'MAE: {bestmodel[3]:.4f}, RMSE: {np.sqrt(bestmodel[2]):.4f}')
print(f'--bestmxwords: {bestmxwords}', f'--bestmxlen: {bestmxlen}', f'--bestembdim: {bestembdim}', f'--bestunit: {bestunit}', sep='\t')

# 2. 베댓만

## 2-1. Dense 층 없이

X2 = df['best']

X2 = X2.apply(okt_tokenizer).values

bestmodel = None
bestloss = 9999
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'mse', 'mae'])

for mxwords in tqdm(range(10, 200+1, 10)):
    for mxlen in [20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            model = train_to_evaluate(X1, y, 'embedding2.h5', mxwords, mxlen, embdim)
            result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, model[2], model[3]]
            if bestloss > model[2]:
                bestmodel = model
                bestloss = model[2]
                bestmxwords = mxwords
                bestmxlen = mxlen
                bestembdim = embdim

print(f'MAE: {bestmodel[3]:.4f}, RMSE: {np.sqrt(bestmodel[2]):.4f}')
print(f'--bestmxwords: {bestmxwords}', f'--bestmxlen: {bestmxlen}', f'--bestembdim: {bestembdim}', sep='\t')

result_df.sort_values(by=['mse']).head(10)

result_df.sort_values(by=['mae']).head(10)

## 특문 제거

X2_a = df['best'].apply(hangul)
X2_a = X2_a.apply(okt_tokenizer).values

from tqdm.auto import tqdm

bestmodel = None
bestloss = 9999
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'mse', 'mae'])

for mxwords in tqdm(range(10, 200+1, 10)):
    for mxlen in [20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            model = train_to_evaluate(X2, y, 'embedding2a.h5', mxwords, mxlen, embdim)
            result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, model[2], model[3]]
            if bestloss > model[2]:
                bestmodel = model
                bestloss = model[2]
                bestmxwords = mxwords
                bestmxlen = mxlen
                bestembdim = embdim

print(f'MAE: {bestmodel[3]:.4f}, RMSE: {np.sqrt(bestmodel[2]):.4f}')
print(f'--bestmxwords: {bestmxwords}', f'--bestmxlen: {bestmxlen}', f'--bestembdim: {bestembdim}', sep='\t')

result_df.sort_values(by=['mse']).head(10)

result_df.sort_values(by=['mae']).head(10)

## 2-2. Dense 층 추가

bestmodel = None
bestloss = 9999
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'unit', 'mse', 'mae'])

for mxwords in tqdm(range(20, 60+1, 10)):
    for mxlen in [20, 30]:
        for embdim in range(10, 100+1, 5):
            for unit in [16, 32, 48, 64]:
                model = train_to_evaluate(X2, y, 'embedding2b.h5', mxwords, mxlen, embdim, unit)
                result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, unit, model[2], model[3]]
                if bestloss > model[2]:
                    bestmodel = model
                    bestloss = model[2]
                    bestmxwords = mxwords
                    bestmxlen = mxlen
                    bestembdim = embdim
                    bestunit = unit

print(f'MAE: {bestmodel[3]:.4f}, RMSE: {np.sqrt(bestmodel[2]):.4f}')
print(f'--bestmxwords: {bestmxwords}', f'--bestmxlen: {bestmxlen}', f'--bestembdim: {bestembdim}', f'--bestunit: {bestunit}', sep='\t')

result_df.sort_values(by=['mse']).head(10)

result_df.sort_values(by=['mae']).head(10)

## 특문 제거

bestmodel = None
bestloss = 9999
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'unit', 'mse', 'mae'])

for mxwords in tqdm(range(20, 60+1, 10)):
    for mxlen in [10, 15, 20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            for unit in [16, 32, 48, 64]:
                model = train_to_evaluate(X2_a, y, 'embedding2c.h5', mxwords, mxlen, embdim, unit)
                result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, unit, model[2], model[3]]
                if bestloss > model[2]:
                    bestmodel = model
                    bestloss = model[2]
                    bestmxwords = mxwords
                    bestmxlen = mxlen
                    bestembdim = embdim
                    bestunit = unit

print(f'MAE: {bestmodel[3]:.4f}, RMSE: {np.sqrt(bestmodel[2]):.4f}')
print(f'--bestmxwords: {bestmxwords}', f'--bestmxlen: {bestmxlen}', f'--bestembdim: {bestembdim}', f'--bestunit: {bestunit}', sep='\t')

result_df.sort_values(by=['mse']).head(10)

result_df.sort_values(by=['mae']).head(10)

# 테스트

'''
MAE: 0.2563, RMSE: 0.4193
--bestmxwords: 30	--bestmxlen: 20	--bestembdim: 15
'''
max_words = 30
maxlen = 20

from tensorflow.keras.models import load_model

X = df['best']

X = X.apply(okt_tokenizer).values

tokenizer = Tokenizer(num_words=max_words)
tokenizer.fit_on_texts(X)

y = df['star']

model = train_to_evaluate(X, y, 'embedding2.h5', 30, 20, 15)

X_seq = tokenizer.texts_to_sequences([okt_tokenizer('와 진짜 재밌음 작가님 기대하고 있을게요')])

X_pad = pad_sequences(X_seq, maxlen=maxlen)

model = load_model('data/embedding2.h5')
pred = model.predict(X_pad)

pred

X_seq = tokenizer.texts_to_sequences([okt_tokenizer('와 진짜 재미없네 이 작가 작품 다 별로임')])

X_pad = pad_sequences(X_seq, maxlen=maxlen)
pred = model.predict(X_pad)
pred

X_seq = tokenizer.texts_to_sequences([okt_tokenizer('이거 뭐냐 왤케 잼씀 ㅋㅋㅋ')])

X_pad = pad_sequences(X_seq, maxlen=maxlen)
pred = model.predict(X_pad)
pred

성능은 괜찮아 보여도 실제로 넣어보면 아닌 것 같다. 아무래도 중앙값이 9.92점으로 높은 점수에 편중되어 있는 y값이라 그런 듯 하다.  
차라리 점수 예측 대신에 임의로 점수를 카테고리화 하여 분류 예측을 해보는 것이 나을 수도 있을 것 같다.
