==> Score/03_MODEL3_EMBEDDING_Classification.ipynb <==
# 필요 라이브러리 호출
import numpy as np
import pandas as pd
from konlpy.tag import Okt
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.preprocessing.text import Tokenizer
import warnings
warnings.filterwarnings('ignore')

# 데이터 불러오기
df = pd.read_csv('data/naver_comment.csv', encoding='utf8')
df.head()

okt = Okt()
def okt_tokenizer(text):
    tokens_ko = okt.morphs(text, stem=True)
    return tokens_ko

import re

def hangul(text):
    return re.sub("[^가-힣ㄱ-하-ㅣ\\s]", "", text)

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Flatten, Dense, Embedding
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint

def train_to_evaluate(X, y, file_name, max_words=10000, maxlen=30, embedding_dim=100, *args):
    
    tokenizer = Tokenizer(num_words=max_words)
    tokenizer.fit_on_texts(X)
    X_seq = tokenizer.texts_to_sequences(X)

    X_pad = pad_sequences(X_seq, maxlen=maxlen)
    y = np.asarray(y)
    
    X_train, X_test, y_train, y_test = train_test_split(X_pad, y, test_size=0.2, random_state=0)
    X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)
    
    model = Sequential()
    model.add(Embedding(max_words, embedding_dim, input_length=maxlen))
    model.add(Flatten())
    for unit in args:
        model.add(Dense(unit, activation='relu'))
    model.add(Dense(1, activation='sigmoid'))
    
    model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
    
    cp = ModelCheckpoint(filepath='data/'+file_name,
                         monitor='accuracy',
                         save_best_only=True)
    
    es = EarlyStopping(monitor='accuracy',
                       patience=15)
    
    
    hist = model.fit(X_train, y_train,
                     epochs=100, batch_size=32,
                     validation_data=(X_val, y_val),
                     callbacks=[cp, es],
                     verbose=0)
    
    
    model.load_weights('data/'+file_name)
    loss, accuracy = model.evaluate(X_val, y_val, verbose=0)
#     print('MAE :', np.round(mae, 4), 'RMSE :', np.round(np.sqrt(mse), 4))
        
    return model, hist, loss, accuracy

# 하위 25%=9.77점이므로, 9.8점 이하면 0, 9.8점 초과면 1로 이진 분류
def category(star):
    if star > 9.8:
        return 1
    else:
        return 0

# 베댓만

## 1. Dense 층 없이

X1 = df['best']
y = df['star'].apply(category)

X1 = X1.apply(okt_tokenizer).values

from tqdm import tqdm

bestmodel = None
bestacc = 0
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'loss', 'accuracy'])

for mxwords in tqdm(range(10, 100+1, 10)):
    for mxlen in [20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            model = train_to_evaluate(X1, y, 'embedding1_category.h5', mxwords, mxlen, embdim)
            result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, model[2], model[3]]
            if bestacc < model[3]:
                bestmodel = model
                bestacc = model[3]
                bestmxwords = mxwords
                bestmxlen = mxlen
                bestembdim = embdim
                
bestmodel[0].save('embedding1_category.h5')

result_df.sort_values(by=['loss']).head(10)

result_df.sort_values(by=['accuracy'], ascending=False).head(10)

## 특문 제거

X1a = df['best'].apply(hangul)
X1a = X1a.apply(okt_tokenizer).values

bestmodel = None
bestacc = 0
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'loss', 'accuracy'])

for mxwords in tqdm(range(10, 100+1, 10)):
    for mxlen in [20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            model = train_to_evaluate(X1a, y, 'embedding1a_category.h5', mxwords, mxlen, embdim)
            result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, model[2], model[3]]
            if bestacc < model[3]:
                bestmodel = model
                bestacc = model[3]
                bestmxwords = mxwords
                bestmxlen = mxlen
                bestembdim = embdim
                
bestmodel[0].save('embedding1a_category.h5')

print(f'Accuracy: {bestmodel[3]:.4f}, loss: {bestmodel[2]:.4f}')
print(f'--bestmxwords: {bestmxwords}, --bestmxlen: {bestmxlen}, --bestembdim: {bestembdim}')

result_df.sort_values(by=['loss']).head(10)

result_df.sort_values(by=['accuracy'], ascending=False).head(10)

## 2. Dense 층 추가

bestmodel = None
bestacc = 0
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'unit', 'loss', 'accuracy'])

for mxwords in tqdm(range(20, 60+1, 10)):
    for mxlen in [10, 15, 20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            for unit in [16, 32, 48, 64]:
                model = train_to_evaluate(X1, y, 'embedding1b_category.h5', mxwords, mxlen, embdim, unit)
                result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, unit, model[2], model[3]]
                if bestacc < model[3]:
                    bestmodel = model
                    bestacc = model[3]
                    bestmxwords = mxwords
                    bestmxlen = mxlen
                    bestembdim = embdim
                    bestunit = unit
                
bestmodel[0].save('embedding1b_category.h5')

print(f'Accuracy: {bestmodel[3]:.4f}, loss: {bestmodel[2]:.4f}')
print(f'--bestmxwords: {bestmxwords}, --bestmxlen: {bestmxlen}, --bestembdim: {bestembdim}, --bestunit: {bestunit}')

result_df.sort_values(by=['loss']).head(10)

result_df.sort_values(by=['accuracy'], ascending=False).head(10)

## 특문 제거

bestmodel = None
bestacc = 0
bestmxwords = None
bestmxlen = None
bestembdim = None
bestunit = None
result_df = pd.DataFrame(columns=['maxwords', 'maxlen', 'embedding_dim', 'unit', 'loss', 'accuracy'])

for mxwords in tqdm(range(20, 60+1, 10)):
    for mxlen in [10, 15, 20, 25, 30]:
        for embdim in range(10, 100+1, 5):
            for unit in [16, 32, 48, 64]:
                model = train_to_evaluate(X1a, y, 'embedding1c_category.h5', mxwords, mxlen, embdim, unit)
                result_df.loc[len(result_df)] = [mxwords, mxlen, embdim, unit, model[2], model[3]]
                if bestacc < model[3]:
                    bestmodel = model
                    bestacc = model[3]
                    bestmxwords = mxwords
                    bestmxlen = mxlen
                    bestembdim = embdim
                    bestunit = unit
                
bestmodel[0].save('embedding1c_category.h5')

print(f'Accuracy: {bestmodel[3]:.4f}, loss: {bestmodel[2]:.4f}')
print(f'--bestmxwords: {bestmxwords}, --bestmxlen: {bestmxlen}, --bestembdim: {bestembdim}, --bestunit: {bestunit}')

result_df.sort_values(by=['loss']).head(10)

result_df.sort_values(by=['accuracy'], ascending=False).head(10)

# 테스트

'''
Accuracy: 0.7398, loss: 1.0082
--bestmxwords: 50, --bestmxlen: 25, --bestembdim: 10, --bestunit: 32
'''
max_words = 50
maxlen = 25

from tensorflow.keras.models import load_model

tokenizer = Tokenizer(num_words=max_words)
tokenizer.fit_on_texts(X1a)
X_seq = tokenizer.texts_to_sequences(X1a)
X_pad = pad_sequences(X_seq, maxlen=maxlen)
y = np.asarray(y)

X_train, X_test, y_train, y_test = train_test_split(X_pad, y, test_size=0.2, random_state=0)
X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)

model = load_model('embedding1c_category.h5')

model.evaluate(X_test, y_test)

X_train_orig, X_test_orig, y_train_orig, y_test_orig = train_test_split(df['best'], df['star'], test_size=0.2, random_state=0)

pred = model.predict(X_test)

y_test_orig[y_test_orig < 9.8]

y_test_orig_copy = y_test_orig.reset_index(drop=True)

y_test_orig_copy[y_test_orig_copy < 9.8]

X_test_orig.iloc[86]

y_test_orig.iloc[86], pred[86]

X_test_orig.iloc[137]

y_test_orig.iloc[137], pred[137]

X_test_orig.iloc[113]

y_test_orig.iloc[113], pred[113]

결과는 별로다.  
데이터 수집 시점에서 타 플랫폼에서 넘어온 작품이 완결로 넘어와서 베댓 수가 적으면서 점수가 낮은 작품 때문에 정확도가 낮은 것 같다.  
또는 3:1 정도로 데이터가 치중되서 그런 것일 수도 있다. (아니면 애초에 댓글로 점수를 예측한다는 가정부터 틀린 것일지도)  
우선 데이터 재수집(작품의 하트(관심도)가 100개 이상인 작품만 수집) + 1:1:1:1로 4등분하여 재시도 하는 걸로
